Replaced project file with working drag and drop hello world.
