+++ Android Practice/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Fri Nov 01 09:54:36 EDT 2019
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-5.4.1-all.zip
