--- Android_practice/.idea/.name
@@ -1,1 +1,0 @@
-Practice app
