--- Android_practice/app/src/main/res/values/strings.xml
@@ -1,4 +1,0 @@
-<resources>
-    <string name="app_name">Practice app</string>
-    <string name="action_settings">Settings</string>
-</resources>
