--- Android_practice/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Sun Oct 27 11:14:31 EDT 2019
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-5.4.1-all.zip
