--- Android_practice/settings.gradle
@@ -1,2 +1,0 @@
-include ':app'
-rootProject.name='Practice app'
